:white_check_mark: Refactoring of tests, fixing and adding tests. Changing the structure to make the tests more understandable.
