+++ docker-compose.yml
@@ -1,0 +1,28 @@
+# Generated by Athena v0.1.0 from MISSING_IMAGE_TEST deployment
+# Generated: 2025-09-16 16:56:36 UTC
+# Features: Intelligent defaults, optimized networking, enhanced health checks
+# DO NOT EDIT MANUALLY - This file is auto-generated
+
+# Services: 1 configured with intelligent defaults
+
+services:
+  test_service:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    container_name: missing-image-test-test_service
+    ports:
+    - 8080:80
+    restart: unless-stopped
+    networks:
+    - missing_image_test_network
+    pull_policy: missing
+    labels:
+      athena.project: MISSING_IMAGE_TEST
+      athena.type: generic
+      athena.service: test_service
+      athena.generated: 2025-09-16
+networks:
+  missing_image_test_network:
+    driver: bridge
+name: MISSING_IMAGE_TEST
