+++ tests/integration/mod.rs
@@ -2,5 +2,5 @@
 pub mod cli_commands_test;
 pub mod docker_compose_generation_test;
 pub mod error_handling_test;
-pub mod boilerplate_generation_test;
+pub mod boilerplate;
 pub mod structural;
